Update instrument config panel, using composition API to allow users to create a custom form
