+++ package.json
@@ -42,6 +42,7 @@
     "vue-template-compiler": "^2.6.11"
   },
   "peerDependencies": {
+    "@vue/composition-api": "^1.0.0-rc.7",
     "bootstrap": "^4.5.x",
     "bootstrap-vue": "^2.20.x",
     "jquery": "^3.5.x",
