+++ src/components/RequestGroupComposition/RequestGroup.vue
@@ -118,6 +118,9 @@
         <template #instrument-config-help="data">
           <slot name="instrument-config-help" :data="data.data"></slot>
         </template>
+        <template #instrument-config-form="data">
+          <slot name="instrument-config-form" :data="data.data"></slot>
+        </template>
         <template #target-help="data">
           <slot name="target-help" :data="data.data"></slot>
         </template>
