+++ src/composables/index.js
@@ -1,0 +1,3 @@
+import baseInstrumentConfigPanel from './baseInstrumentConfigPanel.js';
+
+export { baseInstrumentConfigPanel };
